# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuard Policy Engine API
  description: |
    **AgentGuard** is a runtime security platform for AI agents. It provides policy-based evaluation,
    audit trails, PII detection, MCP server governance, agent hierarchy management, and OWASP compliance reporting.

    ## Authentication

    AgentGuard uses three authentication schemes:

    - **API Key** (`x-api-key` header): Your tenant API key, prefixed `ag_live_`. Used for most authenticated endpoints.
    - **Agent Key** (`x-api-key` header): Agent-scoped key, prefixed `ag_agent_`. Used for evaluate calls from specific agents.
    - **Admin Key** (`x-admin-key` header or `x-api-key`): Platform admin key. Required for admin-only endpoints.

    ## Rate Limits

    - Unauthenticated: 10 requests/min per IP
    - Authenticated: 100 requests/min per IP
    - Signup: 5 per hour per IP

    ## Base URL

    Production: `https://api.agentguard.tech/api/v1`
  version: 0.10.0
  contact:
    name: AgentGuard Support
    url: https://agentguard.tech
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agentguard.tech/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Tenant API key (prefixed `ag_live_`) or agent key (prefixed `ag_agent_`)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Platform admin key (set via ADMIN_KEY environment variable)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token (future use)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid API key"
        details:
          type: array
          items:
            type: string
          description: Additional error details (validation errors)
...
paths:
...
  /license/issue:
    post:
      tags: [License]
      summary: Issue license key (admin)
      description: |
        Issue a new enterprise license key for a tenant.
        The raw `AGKEY-...` string is returned **once** — only its SHA-256 hash is stored.
        **Requires admin key.**
      operationId: issueLicense
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, tier]
              properties:
                tenantId:
                  type: string
                  format: uuid
                  description: Target tenant ID
                tier:
                  type: string
                  enum: [free, pro, enterprise]
                features:
                  type: array
                  items:
                    type: string
                  description: Optional feature overrides
                customLimits:
                  type: object
                  description: Custom limits (overrides tier defaults)
                expiresAt:
                  type: string
                  format: date-time
                  description: License expiry date (null=no expiry)
                stripeSubscriptionId:
                  type: string
                  description: Stripe subscription ID for billing linkage
                metadata:
                  type: object
      responses:
        '201':
          description: License issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  licenseKey:
                    type: string
                    description: "Raw license key (AGKEY-...) — shown once only"
                  tenantId:
                    type: string
                  tier:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Admin key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /license/status:
    get:
      tags: [License]
      summary: Get license status
      description: Returns the current license tier, feature flags, limits, and this month's usage.
      operationId: getLicenseStatus
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: License status
          content:
            application/json:
              schema:
                type: object
                properties:
                  tier:
                    type: string
                    enum: [free, pro, enterprise]
                  status:
                    type: string
                    enum: [active, expired, revoked, none]
                  features:
                    type: array
                    items:
                      type: string
                  limits:
                    type: object
                    description: Feature-specific limits for this tier
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
                  usage:
                    type: object
                    description: Current month usage counters
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /license/validate:
    post:
      tags: [License]
      summary: Validate license key (public)
      description: |
        Validate a license key against the certificate revocation list (CRL).
        Public endpoint — no authentication required.
      operationId: validateLicense
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [licenseKey]
              properties:
                licenseKey:
                  type: string
                  description: The AGKEY-... license key to validate
                  example: "AGKEY-a1b2c3d4..."
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  tier:
                    type: string
                    nullable: true
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
                  revoked:
                    type: boolean
        '400':
          description: Missing or invalid license key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /license/revoke:
    post:
      tags: [License]
      summary: Revoke license key (admin)
      description: |
        Revoke a tenant's license key, adding it to the CRL.
        The tenant immediately loses access to paid features.
        **Requires admin key.**
      operationId: revokeLicense
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId:
                  type: string
                  description: Tenant whose license to revoke
                reason:
                  type: string
                  description: Reason for revocation
      responses:
        '200':
          description: License revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean
                  tenantId:
                    type: string
        '401':
          description: Admin key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No active license for tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /license/usage:
    get:
      tags: [License]
      summary: License usage history
      description: Monthly usage history for the last 12 months (evaluations, blocked calls, etc.)
      operationId: getLicenseUsage
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Monthly usage history
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: string
                          description: "Month in YYYY-MM format"
                          example: "2024-01"
                        evaluations:
                          type: integer
                        blocked:
                          type: integer
                        costCents:
                          type: number
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── PLAYGROUND (additional) ───────────────────────────────────────────────
...